apps/live-agent/pyproject.toml
[project]
name = "reptrainer-live-agent"
version = "0.1.0"
description = "Python ADK live session service for Reptrainer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "google-adk[vertexai]>=1.20.0",
    "google-cloud-firestore>=2.19",
    "firebase-admin>=6.5",
    "google-genai>=1.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.setuptools]
package-dir = {"" = "app"}
packages = ["persona_agent", "tools"]
py-modules = ["auth", "config", "firestore_client", "main", "vertex_search", "analysis"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
